move gossip_public_key fn
