+++ dkg-gadget/src/non_dkg_message.rs
@@ -1,12 +1,13 @@
 // Handles non-dkg messages
 
-use crate::{worker::DKGWorker, Client};
+use crate::{types::dkg_topic, worker::DKGWorker, Client};
+use codec::Encode;
 use dkg_primitives::{
 	crypto::Public,
-	types::{DKGError, DKGMessage, DKGMsgPayload},
+	types::{DKGError, DKGMessage, DKGMsgPayload, DKGPublicKeyMessage, RoundId, SignedDKGMessage},
 };
 use dkg_runtime_primitives::{crypto::AuthorityId, AggregatedPublicKeys, DKGApi};
-use log::debug;
+use log::{debug, error, trace};
 use sc_client_api::Backend;
 use sp_runtime::{
 	generic::BlockId,
@@ -90,4 +91,73 @@ where
 	}
 
 	Ok(())
+}
+
+pub(crate) fn gossip_public_key<B, C, BE>(
+	mut dkg_worker: &mut DKGWorker<B, C, BE>,
+	public_key: Vec<u8>,
+	round_id: RoundId,
+) where
+	B: Block,
+	BE: Backend<B>,
+	C: Client<B, BE>,
+	C::Api: DKGApi<B, AuthorityId, <<B as Block>::Header as Header>::Number>,
+{
+	let sr25519_public = dkg_worker
+		.key_store
+		.sr25519_authority_id(&dkg_worker.key_store.sr25519_public_keys().unwrap_or_default())
+		.unwrap_or_else(|| panic!("Could not find sr25519 key in keystore"));
+
+	let public = dkg_worker
+		.key_store
+		.authority_id(&dkg_worker.key_store.public_keys().unwrap_or_default())
+		.unwrap_or_else(|| panic!("Could not find an ecdsa key in keystore"));
+
+	if let Ok(signature) = dkg_worker.key_store.sr25519_sign(&sr25519_public, &public_key) {
+		let encoded_signature = signature.encode();
+		let payload = DKGMsgPayload::PublicKeyBroadcast(DKGPublicKeyMessage {
+			round_id,
+			pub_key: public_key.clone(),
+			signature: encoded_signature.clone(),
+		});
+
+		let message = DKGMessage::<AuthorityId> { id: public.clone(), round_id, payload };
+		let encoded_dkg_message = message.encode();
+
+		match dkg_worker.key_store.sr25519_sign(&sr25519_public, &encoded_dkg_message) {
+			Ok(sig) => {
+				let signed_dkg_message =
+					SignedDKGMessage { msg: message, signature: Some(sig.encode()) };
+				let encoded_signed_dkg_message = signed_dkg_message.encode();
+
+				dkg_worker.gossip_engine.lock().gossip_message(
+					dkg_topic::<B>(),
+					encoded_signed_dkg_message.clone(),
+					true,
+				);
+			},
+			Err(e) => trace!(
+				target: "dkg",
+				"🕸️  Error signing DKG message: {:?}",
+				e
+			),
+		}
+
+		let mut aggregated_public_keys =
+			if dkg_worker.aggregated_public_keys.get(&round_id).is_some() {
+				dkg_worker.aggregated_public_keys.get(&round_id).unwrap().clone()
+			} else {
+				AggregatedPublicKeys::default()
+			};
+
+		aggregated_public_keys
+			.keys_and_signatures
+			.push((public_key.clone(), encoded_signature));
+
+		dkg_worker.aggregated_public_keys.insert(round_id, aggregated_public_keys);
+
+		debug!(target: "dkg", "Gossiping local node  {:?} public key and signature", public)
+	} else {
+		error!(target: "dkg", "Could not sign public key");
+	}
 }
